requirements.txt
Django==2.0.5
psycopg2-binary==2.7.4
pytz==2018.4
python-dotenv==0.8.2